+++ rag/store/chromem/store_test.go
@@ -33,7 +33,7 @@ func TestChromemStore(t *testing.T) {
 			Metadata: map[string]interface{}{
 				"category": "fruit",
 			},
-			Embedding: []float64{1.0, 0.0, 0.0},
+			Embedding: []float32{1.0, 0.0, 0.0},
 		},
 		{
 			ID:   "2",
@@ -42,7 +42,7 @@ func TestChromemStore(t *testing.T) {
 			Metadata: map[string]interface{}{
 				"category": "vehicle",
 			},
-			Embedding: []float64{0.0, 1.0, 0.0},
+			Embedding: []float32{0.0, 1.0, 0.0},
 		},
 	}
 
@@ -51,7 +51,7 @@ func TestChromemStore(t *testing.T) {
 	assert.Len(t, ids, 2)
 
 	// 3. Query (Exact match for Apple)
-	queryVec := []float64{1.0, 0.0, 0.0}
+	queryVec := []float32{1.0, 0.0, 0.0}
 	query := schema.VectorStoreQuery{
 		Embedding: queryVec,
 		TopK:      1,
@@ -67,7 +67,7 @@ func TestChromemStore(t *testing.T) {
 	assert.InDelta(t, 1.0, results[0].Score, 0.0001)
 
 	// 4. Query (Exact match for Car)
-	queryVecCar := []float64{0.0, 1.0, 0.0}
+	queryVecCar := []float32{0.0, 1.0, 0.0}
 	queryCar := schema.VectorStoreQuery{
 		Embedding: queryVecCar,
 		TopK:      1,
@@ -96,13 +96,13 @@ func TestChromemStore_InMemory(t *testing.T) {
 	store, err := NewSimpleChromemStore("", "mem-collection")
 	require.NoError(t, err)
 
-	nodes := []schema.Node{{ID: "A", Text: "Alpha", Embedding: []float64{0.5}}}
+	nodes := []schema.Node{{ID: "A", Text: "Alpha", Embedding: []float32{0.5}}}
 
 	_, err = store.Add(ctx, nodes)
 	require.NoError(t, err)
 
 	query := schema.VectorStoreQuery{
-		Embedding: []float64{0.5},
+		Embedding: []float32{0.5},
 		TopK:      1,
 	}
 	res, err := store.Query(ctx, query)
